refactor(sync): inline createTransport into createSyncExtension
createTransport had exactly one consumer and existed solely to separate
'transport' from 'extension' — a boundary that created more indirection
than clarity. Every new protocol feature (SYNC_STATUS, RPC) required
threading behavior through callbacks instead of writing it inline.

Inlining eliminates:
- Transport, TransportConfig, MessageHandler, MessageHandlerContext types
- onRpcMessage callback (RPC switch case now accesses pendingRequests directly)
- messageHandlers config + CORE_MESSAGE_TYPES guard (zero consumers)
- The status-change listener for clearPendingRequests (called directly
  in the supervisor loop before each connection attempt)
- websocket-transport.ts as a file

The three module-level helpers (createBackoff, createLivenessMonitor,
createStatusEmitter) survive — they're genuinely reusable state machines.

Transport tests need rewriting to test through createSyncExtension.
